requirements.txt
ccxt==2.4.88
pandas==2.0.2
python-telegram-bot==20.3
tabulate==0.9.0
telethon==1.33.1
supervisor==4.2.5
python-binance==1.0.17
mysql-connector-python==8.2.0
anyio==4.2.0
cryptg==0.4.0
numpy==1.24.3
joblib==1.3.2
xgboost==1.7.6
scikit-learn==1.3.0